Dump user agent emulation preference; we're doing mode only.
It's too complicated to make a proper UA emulation for a bunch
of browsers. Especially if you want it up to date instead of just
picking a random version.

Plus all I really give a hoot about for my own usage is being able
to force a desktop site when necessarily.